Add per-project PLANNER dashboard; move description + hyperGLANCE out of the project form (B)
Every project card gains a PLANNER button opening a planning dashboard
themed to the project's effective color — bottom sheet on mobile,
centered modal on desktop.

The planner hosts:
- Notes (the project description), autosaved on blur/close. Moved out
  of the New/Edit Project form along with hyperGLANCE, which slims that
  form down to identity fields (title, goal, color, users, status).
- Scheduled / Unscheduled task columns (stacked on mobile) reusing the
  SCHED task cards, with day grouping in the scheduled column,
  completed tasks sinking to the bottom, and a quick-add that inherits
  the project's color and assigned users.
- hyperGLANCE settings via the new self-contained HyperGlanceEditor
  (extracted verbatim from ProjectForm), which emits the assembled
  config on user changes only — mounting the planner can't dirty the
  project — and saves through updateProject immediately.

ProjectForm omits description/hyperglance from its save payload so
existing values are preserved by updateProject's merge. SchedTaskCard
gains an isInbox prop so the planner's unscheduled column edits and
completes inbox tasks correctly.
